manageevent/requirement.txt
djangorestframework==3.15.0
django-cors-headers==4.0.0